+++ src/main/java/com/turn/ttorrent/client/SharedTorrent.java
@@ -316,8 +316,8 @@ public class SharedTorrent extends Torrent implements PeerActivityListener {
 			// length. Let's make sure we get the right piece length in any
 			// situation.
 			long off = ((long)idx) * this.pieceLength;
-			int len = Math.min(
-				(int)(this.bucket.size() - off),
+			long len = Math.min(
+				this.bucket.size() - off,
 				this.pieceLength);
 
 			this.pieces[idx] = new Piece(this.bucket, idx, off, len, hash,
